# ARROW-6803: [Rust] [DataFusion] Performance optimization for single partition aggregate queries