This PR optimizes the case where there is a single partition being aggregated. In this use case there is no need for a secondary aggregation step.

I also updated the aggregate benchmark to run each query using the legacy query execution code as well as the new query execution code. The results are relatively similar now.

```
aggregate_query_no_group_by                                                                             
                        time:   [13.139 us 13.209 us 13.286 us]
                        change: [+0.0634% +1.0384% +2.2937%] (p = 0.07 > 0.05)
                        No change in performance detected.
Found 7 outliers among 100 measurements (7.00%)
  2 (2.00%) high mild
  5 (5.00%) high severe

aggregate_query_no_group_by_new                                                                             
                        time:   [24.168 us 24.199 us 24.231 us]
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  8 (8.00%) high severe

aggregate_query_group_by                                                                             
                        time:   [42.192 us 42.245 us 42.305 us]
                        change: [+0.2058% +1.2808% +2.1179%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) high mild
  8 (8.00%) high severe

aggregate_query_group_by_new                                                                             
                        time:   [48.398 us 48.465 us 48.534 us]
                        change: [+1.5342% +1.8353% +2.1283%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

aggregate_query_group_by_with_filter                                                                            
                        time:   [76.195 us 76.339 us 76.488 us]
                        change: [-2.8609% -0.3000% +1.6410%] (p = 0.84 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  8 (8.00%) high mild
  1 (1.00%) high severe

aggregate_query_group_by_with_filter_new                                                                            
                        time:   [81.111 us 81.482 us 81.862 us]
Found 2 outliers among 100 measurements (2.00%)

```